Add initial component
