+++ src/components/card.js
@@ -1,17 +1,12 @@
 import {createCardTemplate} from '../templates/cards';
-import {createElement} from '../utils';
+import Initial from './initial';
 
-export default class Card {
+export default class Card extends Initial {
   constructor(data) {
-    this._data = data;
-    this._element = null;
+    super(data);
 
     this._onClick = null;
     this._onCommentsClick = this._onCommentsClick.bind(this);
-  }
-
-  get element() {
-    return this._element;
   }
 
   get template() {
@@ -26,18 +21,12 @@ export default class Card {
     return typeof this._onClick === `function` && this._onClick();
   }
 
-  _bind() {
+  createListeners() {
     if (this._element) {
       this
         ._element
         .querySelector(`.film-card__comments`)
         .addEventListener(`click`, this._onCommentsClick);
     }
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this._bind();
-    return this._element;
   }
 }
